0.2.0: New auto file creation. Fixed source bugs
Feature: Now there is the option to have vim-osplugin to automatically
create the config files in the directory for you. It if off by default

Bugfix: Fixed a bug where the sourcing of linux and macOS files would
actually open the windows config files

TODO: Whenever we want to automatically create a config file for
windows, make it so that the command prompt does not show up.
TODO: Bugfix on Linux
TODO: Add documentation